add feature gates for gql components in extension
